+++ openbank-audit-service/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [0.17.0](https://github.com/JiRaska/open-bank-oss/compare/audit-service-v0.16.0...audit-service-v0.17.0) (2026-08-20)
+
+
+### Features
+
+* **audit:** add KMS-backed audit anchors ([#5844](https://github.com/JiRaska/open-bank-oss/issues/5844)) ([785df7a](https://github.com/JiRaska/open-bank-oss/commit/785df7a0807e05dc6eab828319abe82f3faa891f))
+
 ## [0.16.0](https://github.com/JiRaska/open-bank-oss/compare/audit-service-v0.15.4...audit-service-v0.16.0) (2026-08-17)